requirements.txt
Pillow
pybtex
staticjinja
scrapy
